+++ photutils/psf/matching/__init__.py
@@ -1,0 +1,8 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This subpackage contains modules and packages to generate kernels for
+matching point spread functions.
+"""
+
+from .fourier import *    # noqa
+from .windows import *    # noqa